func-app/requirements.txt
aenum==3.1.15 
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5 
aiolimiter==1.1.0 
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0 
anytree==2.12.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autograd==1.7.0
azure-common==1.1.28
azure-core==1.31.0
azure-cosmos==4.7.0
azure-identity==1.17.1
azure-kusto-data==4.5.1
azure-search-documents==11.5.1
azure-storage-blob==12.23.0
beartype==0.18.5
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
cramjam==2.8.3
cryptography==43.0.1
cycler==0.12.1
dask-expr==1.1.14
dask==2024.9.0
dask[dataframe]==2024.9.0
datashaper==0.0.49
decorator==5.1.1
deprecation==2.1.0
devtools==0.12.2
diskcache==5.6.3
distro==1.9.0
environs==11.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastparquet==2024.5.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gensim==4.3.3
graspologic-native==1.2.1
graspologic==3.4.1
gremlinpython==3.7.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
hyppo==0.4.0
idna==3.10
ijson==3.3.0
importlib-metadata==8.5.0
isodate==0.6.1
jiter==0.5.0
joblib==1.4.2
json-repair==0.25.3
jsonschema-specifications==2023.12.1
jsonschema==4.23.0
kiwisolver==1.4.7
lancedb==0.11.0
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
markdown-it-py[linkify,plugins]==3.0.0
marshmallow==3.22.0
matplotlib==3.9.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msal-extensions==1.2.0
msal==1.31.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4 
openai==1.46.0 
overrides==7.7.0 
packaging==24.1
pandas==2.2.2
partd==1.4.2
patsy==0.5.6
pillow==10.4.0
portalocker==2.10.1
pot==0.9.4
psutil==6.0.0
py==1.11.0
pyaml-env==1.2.1
pyarrow==15.0.2
pycparser==2.22
pydantic-core==2.23.4
pydantic==2.9.2
pygments==2.18.0
pyjwt[crypto]==2.9.0
pylance==0.15.0
pynndescent==0.5.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306 ## This needs to be commented when deploying to the func app.
pyyaml==6.0.2
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
retry==0.9.2
rich==13.8.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
statsmodels==0.14.3
swifter==1.4.0
tenacity==8.5.0
textual==0.74.0
threadpoolctl==3.5.0
tiktoken==0.7.0
toolz==0.12.1
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.6
urllib3==2.2.3
wrapt==1.16.0
yarl==1.11.1
zipp==3.20.2
azure.functions
future
pandas